[Decomp][Player] Cohesion split: move Player taxi-flight methods into PlayerTaxi.cpp
